Improve many things
This commit goes through and improves a lot and fixes a lot of bugs. I lost track of what since it was so many little things. The biggest thing was that it got rid of the onDone model and switched to using NotificationCenter for updates/creations, so that all loaded table views would be updated, not just the direct parent.

+++ Project/TheBooks/ViewClient.swift
@@ -8,8 +8,12 @@
 
 import UIKit
 
+public let CLIENT_UPDATED_NOTIFICATION = Notification.Name("client updated"),
+	CLIENT_CREATED_NOTIFICATION = Notification.Name("client created")
+
 class ViewClient: UITableViewController {
 	let context = (UIApplication.shared.delegate as! AppDelegate).context
+	let center = NotificationCenter.default
 	
 	@IBOutlet weak var firstNameField: UITextField!
 	@IBOutlet weak var lastNameField: UITextField!
@@ -23,24 +27,26 @@ class ViewClient: UITableViewController {
 	}
 	
 	private var firstName: String {
-		get { return firstNameField.text! }
+		get { return firstNameField.text!.trimmingCharacters(in: .whitespacesAndNewlines) }
 		set { firstNameField.text = newValue }
 	}
 	
 	private var lastName: String {
-		get { return lastNameField.text! }
+		get { return lastNameField.text!.trimmingCharacters(in: .whitespacesAndNewlines) }
 		set { lastNameField.text = newValue }
 	}
 	
 	private var phone: String {
-		get { return phoneField.text! }
+		get { return phoneField.text!.trimmingCharacters(in: .whitespacesAndNewlines) }
 		set { phoneField.text = newValue }
 	}
 	
 	private var email: String {
-		get { return emailField.text! }
+		get { return emailField.text!.trimmingCharacters(in: .whitespacesAndNewlines) }
 		set { emailField.text = newValue }
 	}
+	
+	var initialKey: Character?
 	
 	var initialFirstName = "" {
 		didSet {
@@ -64,18 +70,15 @@ class ViewClient: UITableViewController {
 	}
 	
 	override func viewDidLoad() {
-		if let client = client {
-			initialFirstName = client.firstName
-			initialLastName = client.lastName
-			initialPhone = client.phone
-			initialEmail = client.email
-			
+		if client != nil {
 			navigationItem.title = "View Client"
 			navigationItem.rightBarButtonItem = UIBarButtonItem(barButtonSystemItem: .save, target: self, action: #selector(ViewClient.done))
-			navigationItem.rightBarButtonItem?.isEnabled = false
 		} else {
-			navigationItem.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .cancel, target: self, action: #selector(ViewClient.dismissSelf))
+			navigationItem.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .cancel, target: self, action: #selector(dismiss as () -> ()))
 		}
+		
+		setInitials()
+		revalidate()
 	}
 	
 	override func numberOfSections(in tableView: UITableView) -> Int {
@@ -90,20 +93,30 @@ class ViewClient: UITableViewController {
 			self.presentSheet(
 				("Cancel", .cancel, nil),
 				("Delete Client", .destructive, {_ in
-					self.context.delete(self.client!)
-					
-					try? self.context.save()
-					
-					self.onDone!(self.client!)
-					
-					_ = self.navigationController?.popViewController(animated: true)
+					if self.client!.appointments.count > 0 {
+						self.alert(title: "Cannot Delete Client", message: "You must either delete or move all of this client's appointments before this client can be deleted")
+					} else {
+						self.context.delete(self.client!)
+						
+						try? self.context.save()
+						
+						// Post about the deletion to all listeners here
+						self.center.post(name: CLIENT_UPDATED_NOTIFICATION, object: self)
+						
+						_ = self.navigationController?.popViewController(animated: true)
+					}
 				})
 			)
 		}
 	}
 	
-	/// Must be provided in the presenting view
-	var onDone: ((Client) -> ())?
+	func setInitials() {
+		initialKey = client?.key
+		initialFirstName = client?.firstName ?? ""
+		initialLastName = client?.lastName ?? ""
+		initialPhone = client?.phone ?? ""
+		initialEmail = client?.email ?? ""
+	}
 	
 	@IBAction func revalidate() {
 		navigationItem.rightBarButtonItem!.isEnabled =
@@ -114,13 +127,10 @@ class ViewClient: UITableViewController {
 	}
 	
 	@IBAction func done() {
-		let client: Client
+		view.endEditing(true)
 		
-		if self.client == nil {
-			client = Client.make(for: context)
-		} else {
-			client = self.client!
-		}
+		let isUpdate = self.client != nil
+		let client = self.client ?? Client.make(for: context)
 		
 		client.firstName = firstName
 		client.lastName = lastName
@@ -129,7 +139,15 @@ class ViewClient: UITableViewController {
 		
 		try? context.save()
 		
-		onDone!(client)
+		self.client = client
+		
+		let name = isUpdate ? CLIENT_UPDATED_NOTIFICATION : CLIENT_CREATED_NOTIFICATION
+		
+		// Post about the change to all listeners here
+		center.post(name: name, object: self)
+		
+		setInitials()
+		revalidate()
 	}
 	
 	override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
